+++ src/main/java/it/gov/pagopa/fdr/controller/InternalPspsController.java
@@ -12,8 +12,8 @@ import it.gov.pagopa.fdr.controller.model.payment.request.DeletePaymentRequest;
 import it.gov.pagopa.fdr.controller.model.payment.response.PaginatedPaymentsResponse;
 import it.gov.pagopa.fdr.service.FlowService;
 import it.gov.pagopa.fdr.service.PaymentService;
-import it.gov.pagopa.fdr.service.model.FindFlowsByFiltersArgs;
-import it.gov.pagopa.fdr.service.re.model.FdrActionEnum;
+import it.gov.pagopa.fdr.service.model.arguments.FindFlowsByFiltersArgs;
+import it.gov.pagopa.fdr.service.model.re.FdrActionEnum;
 import it.gov.pagopa.fdr.util.Re;
 import it.gov.pagopa.fdr.util.constant.AppConstant;
 import jakarta.ws.rs.Consumes;
