+++ src/main/java/it/gov/pagopa/fdr/controller/TechnicalSupportController.java
@@ -3,7 +3,7 @@ package it.gov.pagopa.fdr.controller;
 import it.gov.pagopa.fdr.controller.interfaces.ISupportController;
 import it.gov.pagopa.fdr.controller.model.flow.response.PaginatedFlowsBySenderAndReceiverResponse;
 import it.gov.pagopa.fdr.service.TechnicalSupportService;
-import it.gov.pagopa.fdr.service.model.FindPaymentsByFiltersArgs;
+import it.gov.pagopa.fdr.service.model.arguments.FindPaymentsByFiltersArgs;
 import java.time.Instant;
 
 public class TechnicalSupportController implements ISupportController {
