+++ src/main/java/it/gov/pagopa/fdr/service/model/re/AppVersionEnum.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.fdr.service.re.model;
+package it.gov.pagopa.fdr.service.model.re;
 
 public enum AppVersionEnum {
   FDR001,
