+++ src/main/java/it/gov/pagopa/fdr/service/model/re/BlobHttpBody.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.fdr.service.re.model;
+package it.gov.pagopa.fdr.service.model.re;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
