+++ src/main/java/it/gov/pagopa/fdr/service/model/re/FdrStatusEnum.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.fdr.service.re.model;
+package it.gov.pagopa.fdr.service.model.re;
 
 public enum FdrStatusEnum {
   CREATED,
