+++ src/main/java/it/gov/pagopa/fdr/service/model/re/HttpTypeEnum.java
@@ -1,0 +1,6 @@
+package it.gov.pagopa.fdr.service.model.re;
+
+public enum HttpTypeEnum {
+  REQ,
+  RES
+}
