+++ src/main/java/it/gov/pagopa/fdr/service/model/re/ReInternal.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.fdr.service.re.model;
+package it.gov.pagopa.fdr.service.model.re;
 
 import lombok.Data;
 import lombok.EqualsAndHashCode;
@@ -10,6 +10,7 @@ import lombok.experimental.SuperBuilder;
 @ToString(callSuper = true)
 @EqualsAndHashCode(callSuper = true)
 public class ReInternal extends ReAbstract {
+
   private boolean fdrPhysicalDelete;
   private FdrStatusEnum fdrStatus;
   private Long revision;
