--- src/main/java/it/gov/pagopa/fdr/service/re/model/HttpTypeEnum.java
@@ -1,6 +1,0 @@
-package it.gov.pagopa.fdr.service.re.model;
-
-public enum HttpTypeEnum {
-  REQ,
-  RES
-}
