+++ src/main/java/it/gov/pagopa/fdr/util/Re.java
@@ -1,6 +1,6 @@
 package it.gov.pagopa.fdr.util;
 
-import it.gov.pagopa.fdr.service.re.model.FdrActionEnum;
+import it.gov.pagopa.fdr.service.model.re.FdrActionEnum;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
@@ -9,5 +9,6 @@ import java.lang.annotation.Target;
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.METHOD)
 public @interface Re {
+
   FdrActionEnum action();
 }
