+++ src/main/resources/application.properties
@@ -85,98 +85,6 @@ adapter.api_config_cache.url=${ADAPTER_API_CONFIG_CACHE_URL:${quarkus.mockserver
 # quarkus.scheduler.cron-type=unix
 api_config_cache.cron.expr=${ADAPTER_API_CONFIG_CACHE_CRON:*/15 * * * * ?}
 ###################
-## QUEUE CONVERSION
-###################
-queue.conversion.enabled=true
-%dev.queue.conversion.enabled=false
-%test.queue.conversion.enabled=false
-#%openapi.queue.conversion.enabled=false
-%openapi_internal.queue.conversion.enabled=false
-%openapi_psp.queue.conversion.enabled=false
-%openapi_organization.queue.conversion.enabled=false
-queue.conversion.connect-str=${QUEUE_CONVERSION_CONNECTION_STRING:${mockserver.azurite.connection-string}}
-%dev.queue.conversion.connect-str=DefaultEndpointsProtocol=http;AccountName=devstoreaccount1;AccountKey=Eby8vdM02xNOcqFlqUwJPLlmEtlCDXJ1OUzFT50uSRZ6IFsuFq2UVErCz4I6tq/K1SZFPTOtr/KBHBeksoGMGw==;BlobEndpoint=http://127.0.0.1:10000/devstoreaccount1;QueueEndpoint=http://127.0.0.1:10001/devstoreaccount1;TableEndpoint=http://127.0.0.1:10002/devstoreaccount1;
-#%openapi.queue.conversion.connect-str=na
-%openapi_internal.queue.conversion.connect-str=na
-%openapi_psp.queue.conversion.connect-str=na
-%openapi_organization.queue.conversion.connect-str=na
-queue.conversion.name=${QUEUE_CONVERSION_NAME:${mockserver.azurite.queue-name}}
-%dev.queue.conversion.name=queueconversion
-#%openapi.queue.conversion.name=na
-%openapi_internal.queue.conversion.name=na
-%openapi_psp.queue.conversion.name=na
-%openapi_organization.queue.conversion.name=na
-###################
-## EVENT HUB RE
-###################
-eHub.re.enabled=true
-%dev.eHub.re.enabled=false
-%test.eHub.re.enabled=false
-#%openapi.eHub.re.enabled=false
-%openapi_internal.eHub.re.enabled=false
-%openapi_psp.eHub.re.enabled=false
-%openapi_organization.eHub.re.enabled=false
-#per il run inserire l'env nel proprio profilo
-ehub.re.connect-str=${EVENT_HUB_RE_CONNECTION_STRING:na}
-%test.ehub.re.connect-str=na
-#%openapi.ehub.re.connect-str=na
-%openapi_internal.ehub.re.connect-str=na
-%openapi_psp.ehub.re.connect-str=na
-%openapi_organization.ehub.re.connect-str=na
-ehub.re.name=${EVENT_HUB_RE_NAME:fdr-re}
-%dev.ehub.re.name=fdr-re
-#%openapi.ehub.re.name=na
-%openapi_internal.ehub.re.name=na
-%openapi_psp.ehub.re.name=na
-%openapi_organization.ehub.re.name=na
-blob.re.connect-str=${BLOB_RE_CONNECTION_STRING:${mockserver.azurite.connection-string}}
-%dev.blob.re.connect-str=DefaultEndpointsProtocol=http;AccountName=devstoreaccount1;AccountKey=Eby8vdM02xNOcqFlqUwJPLlmEtlCDXJ1OUzFT50uSRZ6IFsuFq2UVErCz4I6tq/K1SZFPTOtr/KBHBeksoGMGw==;BlobEndpoint=http://127.0.0.1:10000/devstoreaccount1;QueueEndpoint=http://127.0.0.1:10001/devstoreaccount1;TableEndpoint=http://127.0.0.1:10002/devstoreaccount1;
-#%openapi.blob.re.connect-str=na
-%openapi_internal.blob.re.connect-str=na
-%openapi_psp.blob.re.connect-str=na
-%openapi_organization.blob.re.connect-str=na
-blob.re.containername=${BLOB_RE_CONTAINER_NAME:${mockserver.azurite.container-name}}
-%dev.blob.re.containername=blobcontainerre
-#%openapi.blob.re.containername=na
-%openapi_internal.blob.re.containername=na
-%openapi_psp.blob.re.containername=na
-%openapi_organization.blob.re.containername=na
-###################
-## BLOB HISTORY
-###################
-history.enabled=true
-%dev.history.enabled=false
-%test.history.enabled=true
-%openapi_internal.history.enabled=false
-%openapi_psp.history.enabled=false
-%openapi_organization.history.enabled=false
-blob.history.connect-str=${BLOB_HISTORY_CONNECTION_STRING:${mockserver.azurite.connection-string}}
-%dev.blob.history.connect-str=DefaultEndpointsProtocol=http;AccountName=devstoreaccount1;AccountKey=Eby8vdM02xNOcqFlqUwJPLlmEtlCDXJ1OUzFT50uSRZ6IFsuFq2UVErCz4I6tq/K1SZFPTOtr/KBHBeksoGMGw==;BlobEndpoint=http://127.0.0.1:10000/devstoreaccount1;QueueEndpoint=http://127.0.0.1:10001/devstoreaccount1;TableEndpoint=http://127.0.0.1:10002/devstoreaccount1;
-%openapi_internal.blob.history.connect-str=na
-%openapi_psp.blob.history.connect-str=na
-%openapi_organization.blob.history.connect-str=na
-blob.history.containername=${BLOB_HISTORY_CONTAINER_NAME:${mockserver.azurite.history.container-name}}
-%dev.blob.history.containername=fdrhistory
-%openapi_internal.blob.history.containername=na
-%openapi_psp.blob.history.containername=na
-%openapi_organization.blob.history.containername=na
-table.history.connect-str=${TABLE_HISTORY_CONNECTION_STRING:${mockserver.azurite.connection-string}}
-%dev.table.history.connect-str=DefaultEndpointsProtocol=http;AccountName=devstoreaccount1;AccountKey=Eby8vdM02xNOcqFlqUwJPLlmEtlCDXJ1OUzFT50uSRZ6IFsuFq2UVErCz4I6tq/K1SZFPTOtr/KBHBeksoGMGw==;BlobEndpoint=http://127.0.0.1:10000/devstoreaccount1;QueueEndpoint=http://127.0.0.1:10001/devstoreaccount1;TableEndpoint=http://127.0.0.1:10002/devstoreaccount1;
-%openapi_internal.table.history.connect-str=na
-%openapi_psp.table.history.connect-str=na
-%openapi_organization.table.history.connect-str=na
-table.history.tablename.fdrpublish=${TABLE_HISTORY_FDR_PUBLISH_TABLE:${mockserver.azurite.table.fdrpublish}}
-%dev.table.history.tablename.fdrpublish=fdrpublish
-%openapi_internal.table.history.tablename.fdrpublish=na
-%openapi_psp.table.history.tablename.fdrpublish=na
-%openapi_organization.table.history.tablename.fdrpublish=na
-table.history.tablename.fdrpaymentpublish=${TABLE_HISTORY_FDR_PAYMENT_PUBLISH_TABLE:${mockserver.azurite.table.fdrpaymentpublish}}
-%dev.table.history.tablename.fdrpaymentpublish=fdrpaymentpublish
-%openapi_internal.table.history.tablename.fdrpaymentpublish=na
-%openapi_psp.table.history.tablename.fdrpaymentpublish=na
-%openapi_organization.table.history.tablename.fdrpaymentpublish=na
-json.schema.version=V1
-###################
 ## OPENAPI FILTER
 ###################
 #operations.filter=".*"
